Add IListingsManager service
