+++ MediaBrowser.Controller/LiveTv/ILiveTvManager.cs
@@ -36,8 +36,6 @@ namespace MediaBrowser.Controller.LiveTv
         /// <value>The services.</value>
         IReadOnlyList<ILiveTvService> Services { get; }
 
-        IReadOnlyList<IListingsProvider> ListingProviders { get; }
-
         /// <summary>
         /// Gets the new timer defaults asynchronous.
         /// </summary>
@@ -240,41 +238,12 @@ namespace MediaBrowser.Controller.LiveTv
         Task AddInfoToProgramDto(IReadOnlyCollection<(BaseItem Item, BaseItemDto ItemDto)> programs, IReadOnlyList<ItemFields> fields, User user = null);
 
         /// <summary>
-        /// Saves the listing provider.
-        /// </summary>
-        /// <param name="info">The information.</param>
-        /// <param name="validateLogin">if set to <c>true</c> [validate login].</param>
-        /// <param name="validateListings">if set to <c>true</c> [validate listings].</param>
-        /// <returns>Task.</returns>
-        Task<ListingsProviderInfo> SaveListingProvider(ListingsProviderInfo info, bool validateLogin, bool validateListings);
-
-        void DeleteListingsProvider(string id);
-
-        Task<TunerChannelMapping> SetChannelMapping(string providerId, string tunerChannelNumber, string providerChannelNumber);
-
-        TunerChannelMapping GetTunerChannelMapping(ChannelInfo tunerChannel, NameValuePair[] mappings, List<ChannelInfo> providerChannels);
-
-        /// <summary>
-        /// Gets the lineups.
-        /// </summary>
-        /// <param name="providerType">Type of the provider.</param>
-        /// <param name="providerId">The provider identifier.</param>
-        /// <param name="country">The country.</param>
-        /// <param name="location">The location.</param>
-        /// <returns>Task&lt;List&lt;NameIdPair&gt;&gt;.</returns>
-        Task<List<NameIdPair>> GetLineups(string providerType, string providerId, string country, string location);
-
-        /// <summary>
         /// Adds the channel information.
         /// </summary>
         /// <param name="items">The items.</param>
         /// <param name="options">The options.</param>
         /// <param name="user">The user.</param>
         void AddChannelInfo(IReadOnlyCollection<(BaseItemDto ItemDto, LiveTvChannel Channel)> items, DtoOptions options, User user);
-
-        Task<List<ChannelInfo>> GetChannelsForListingsProvider(string id, CancellationToken cancellationToken);
-
-        Task<List<ChannelInfo>> GetChannelsFromListingsProviderData(string id, CancellationToken cancellationToken);
 
         string GetEmbyTvActiveRecordingPath(string id);
