--- MediaBrowser.Controller/LiveTv/TunerChannelMapping.cs
@@ -1,17 +1,0 @@
-#nullable disable
-
-#pragma warning disable CS1591
-
-namespace MediaBrowser.Controller.LiveTv
-{
-    public class TunerChannelMapping
-    {
-        public string Name { get; set; }
-
-        public string ProviderChannelName { get; set; }
-
-        public string ProviderChannelId { get; set; }
-
-        public string Id { get; set; }
-    }
-}
